v0.0.2 update
