+++ src/af_vote/structure_analysis.py
@@ -5,7 +5,8 @@ from typing import List, Optional, Dict, Any
 import numpy as np
 import pandas as pd
 from joblib import Parallel, delayed
-from Bio.PDB import PDBParser, Superimposer
+from Bio.PDB import PDBParser, Superimposer, PPBuilder
+import json
 
 from af_path.sequence_processing import count_sequences_in_a3m
 
@@ -371,3 +372,108 @@ def get_result_df(parent_dir: str,
     results_df = pd.DataFrame(results)
 
     return results_df
+
+
+
+def get_protein_sequence(pdb_filename: str) -> str:
+    """
+    Extract the protein sequence from a PDB file.
+
+    Args:
+        pdb_filename (str): Path to the PDB file.
+
+    Returns:
+        str: The full protein sequence.
+
+    Raises:
+        Exception: If there's an error in processing the PDB file.
+    """
+    try:
+        pdb_parser = PDBParser(QUIET=True)
+        structure = pdb_parser.get_structure("Protein", pdb_filename)
+        ppb = PPBuilder()
+        sequences = []
+        for pp in ppb.build_peptides(structure):
+            sequence = pp.get_sequence()
+            sequences.append(str(sequence))
+
+        full_sequence = ''.join(sequences)
+        return full_sequence
+
+    except Exception as e:
+        logging.error(f"Error getting protein sequence: {e}")
+        raise
+
+def count_sequences_in_a3m(a3m_file: str) -> int:
+    """
+    Count the number of sequences in an A3M file.
+
+    Args:
+        a3m_file (str): Path to the A3M file.
+
+    Returns:
+        int: The number of sequences in the A3M file.
+    """
+    count = 0
+    try:
+        with open(a3m_file, 'r') as file:
+            for line in file:
+                if line.startswith('>'):
+                    count += 1
+    except FileNotFoundError:
+        logging.error(f"A3M file not found: {a3m_file}")
+    return count
+
+def load_filter_modes(file_path: str) -> Dict:
+    """
+    Load filter modes from a JSON file.
+
+    Args:
+        file_path (str): Path to the JSON file containing filter modes.
+
+    Returns:
+        Dict: The loaded filter modes.
+    """
+    with open(file_path, 'r') as f:
+        return json.load(f)
+
+def apply_filters(df_threshold: pd.DataFrame, 
+                  df_operate: pd.DataFrame, 
+                  filter_criteria: List[Dict[str, Any]], 
+                  quantile: float) -> pd.DataFrame:
+    """
+    Apply filters to the dataframe based on specified criteria and quantile thresholds.
+
+    Args:
+        df_threshold (pd.DataFrame): DataFrame used for calculating thresholds.
+        df_operate (pd.DataFrame): DataFrame to apply filters on.
+        filter_criteria (List[Dict[str, Any]]): List of filter criteria dictionaries.
+        quantile (float): Quantile value for threshold calculation.
+
+    Returns:
+        pd.DataFrame: Filtered DataFrame.
+    """
+    filtered_df = df_operate.copy()
+    
+    for criterion in filter_criteria:
+        column_name = criterion['name']
+        method = criterion['method']
+        
+        if method not in ['above', 'below']:
+            logging.warning(f"Invalid filter method: {method}. Skipping this criterion.")
+            continue
+        
+        if column_name not in df_threshold.columns:
+            logging.warning(f"Column {column_name} not found in threshold DataFrame. Skipping this criterion.")
+            continue
+        
+        if method == 'above':
+            threshold_value = df_threshold[column_name].quantile(1 - quantile)
+            filtered_df = filtered_df[filtered_df[column_name] > threshold_value]
+            logging.info(f"Filtering {column_name} above {(1-quantile)*100}% quantile value: {threshold_value}")
+        elif method == 'below':
+            threshold_value = df_threshold[column_name].quantile(quantile)
+            filtered_df = filtered_df[filtered_df[column_name] < threshold_value]
+            logging.info(f"Filtering {column_name} below {quantile*100}% quantile value: {threshold_value}")
+    
+    return filtered_df
